+++ models/Product.ts
@@ -58,7 +58,7 @@ Product.init(
       allowNull: true
     },
     quantity: {
-      type: DataTypes.INTEGER,
+      type: DataTypes.DECIMAL(12, 3),
       allowNull: false,
       defaultValue: 0,
       validate: {
